update: diary 엔티티 totalikes 필드 추가로 인한 좋아요 및 리스트 조회 로직 수정 #144

+++ src/main/java/com/moa/domain/diary/diary/dto/DiaryDto.java
@@ -72,7 +72,7 @@ public class DiaryDto {
         private String diaryTitle;
         private String diaryContents;
         private LocalDate diaryPublishedAt;
-        private Integer totalLikes;
+        private Long totalLikes;
     }
 
 }
